* release/v7.2.0:
  Bump version
  Update documentation
  Move app.url & session.domain runtime override to modules & application layer
  Use the new app.domains config option
  Remove config options and global helpers to modules and application layer
  Check before detaching pages if deleted entity was soft deleted

+++ src/Traits/Tenantable.php
@@ -100,7 +100,8 @@ trait Tenantable
         }
 
         static::deleted(function (self $model) {
-            $model->tenants()->detach();
+            // Check if this is a soft delete or not by checking if `SoftDeletes::isForceDeleting` method exists
+            (method_exists($model, 'isForceDeleting') && ! $model->isForceDeleting()) || $model->tenants()->detach();
         });
     }
